Add some basic components
