+++ frontend/package.json
@@ -10,6 +10,7 @@
   "dependencies": {
     "core-js": "^3.20.3",
     "echarts": "^5.2.2",
+    "sass": "^1.49.0",
     "vue": "^3.0.0",
     "vue-echarts": "^6.0.2"
   },
